Refactor remote signing to its own crate and support rust_native_crypto - use certificate parsing and validation according to crypto provider - updates unit tests - fix cfg for wasm unittests
