+++ sdk/Cargo.toml
@@ -49,8 +49,7 @@ rust_native_crypto = [
     "dep:spki",
 ]
 remote_signing = [
-    "openssl",
-    "dep:reqwest",
+    "dep:reqwest"
 ]
 # TODO: make our async tests always use the async API so we don't need ureq + reqwest
 #       see https://github.com/contentauth/c2pa-rs/issues/1364
