+++ sdk/src/crypto/raw_signature/openssl/signers/mod.rs
@@ -20,9 +20,6 @@ mod ecdsa_signer;
 mod ed25519_signer;
 mod rsa_signer;
 
-#[cfg(feature = "remote_signing")]
-mod remote_signer;
-
 /// Return a built-in [`RawSigner`] instance using the provided signing
 /// certificate and private key.
 ///
@@ -64,29 +61,4 @@ pub(crate) fn signer_from_cert_chain_and_private_key(
             )?,
         )),
     }
-}
-
-/// Return a built-in [`RawSigner`] instance using the provided signing
-/// certificate and url to the remote signing service.
-///
-/// Which signers are available depends on the remote signing service.
-/// It is assumed that the signature returned by the remote signing service
-/// conforms to the public key in the certificate chain
-///
-/// May return an `Err` response if the certificate chain is invalid.
-#[cfg(feature = "remote_signing")]
-pub(crate) fn signer_from_cert_chain_and_url(
-    cert_chain: &[u8],
-    url: url::Url,
-    alg: SigningAlg,
-    time_stamp_service_url: Option<String>,
-) -> Result<Box<dyn RawSigner + Send + Sync>, RawSignerError> {
-    Ok(Box::new(
-        remote_signer::RemoteRawSigner::from_cert_chain_and_url(
-            cert_chain,
-            url,
-            alg,
-            time_stamp_service_url,
-        )?,
-    ))
 }
