+++ sdk/src/crypto/raw_signature/openssl/signers/rsa_signer.rs
@@ -20,10 +20,7 @@ use openssl::{
 
 use crate::crypto::{
     raw_signature::{
-        openssl::{
-            cert_chain::parse_and_check_chain_order,
-            OpenSslMutex,
-        },
+        openssl::{cert_chain::parse_and_check_chain_order, OpenSslMutex},
         RawSigner, RawSignerError, SigningAlg,
     },
     time_stamp::TimeStampProvider,
@@ -56,7 +53,6 @@ impl RsaSigner {
         alg: SigningAlg,
         time_stamp_service_url: Option<String>,
     ) -> Result<Self, RawSignerError> {
-
         //get validly ordered certs in DER format
         let cert_chain = parse_and_check_chain_order(cert_chain)?;
