+++ sdk/src/crypto/raw_signature/rust_native/signers/ed25519_signer.rs
@@ -15,10 +15,11 @@ use ed25519_dalek::{pkcs8::DecodePrivateKey, SigningKey};
 use x509_parser::{error::PEMError, pem::Pem};
 
 use crate::crypto::{
-    raw_signature::{RawSigner, RawSignerError, SigningAlg},
+    raw_signature::{
+        rust_native::cert_chain::cert_chain_to_der, RawSigner, RawSignerError, SigningAlg,
+    },
     time_stamp::TimeStampProvider,
 };
-use crate::crypto::raw_signature::rust_native::cert_chain::cert_chain_to_der;
 
 /// Implements `RawSigner` trait using `ed25519_dalek` crate's implementation of
 /// Edwards Curve encryption.
@@ -39,7 +40,7 @@ impl Ed25519Signer {
         private_key: &[u8],
         time_stamp_service_url: Option<String>,
     ) -> Result<Self, RawSignerError> {
-        let cert_chain =  cert_chain_to_der(cert_chain)?;
+        let cert_chain = cert_chain_to_der(cert_chain)?;
         let cert_chain_len = cert_chain.iter().fold(0usize, |sum, c| sum + c.len());
 
         let private_key_pem = std::str::from_utf8(private_key).map_err(|e| {
